isis2503-201910-s4-sprinters-master/AppMonolítica/requirements.txt
Django
djangorestframework
django_progressive_web_app
psycopg2-binary
request
PyJWT==1.7.1
python-dotenv==0.10.1
social-auth-app-django==3.1.0
cryptography==2.6.1
psycopg2
kafka-python
djangorestframework